Move file formats to `Formats` directory
