+++ src/Formats/Color/ColorFile.h
