+++ src/Formats/Color/ColorFileCommons.cpp
