+++ src/Formats/Color/ColorFileCommons.h
