+++ src/Formats/Color/ColorFileOld.h
