+++ src/Formats/Color/Colors/Color.cpp
