+++ src/Formats/Color/Colors/ColorCommons.cpp
