+++ src/Formats/Color/Colors/WorldColor.cpp
