+++ src/Formats/Color/Colors/WorldColor.h
