+++ src/Formats/Localization/Language.h
@@ -4,8 +4,8 @@
 #ifndef LANGUAGE_H
 #define LANGUAGE_H
 
-#include "../libLCE.h"
-#include "../IO/BinaryIO.h"
+#include "../../libLCE.h"
+#include "../../IO/BinaryIO.h"
 
 #include <string>
 #include <vector>
